*alpha* ABB updates for electrons/protons

+++ us_somo/develop/include/us_hydrodyn_pdbdefs.h
@@ -188,7 +188,7 @@ struct PDB_model
    double hydration;                        // number of waters assigned
    double hydration_gg;                     // hydration g/g
    double molar_volume;                     // mv of model
-   int   num_elect;                         // number of electrons
+   double num_elect;                         // number of electrons
    int   num_SS_bonds;                      // number of SS bonds
    int   num_SH_free;                       // number of free SH
 
@@ -458,9 +458,9 @@ struct hybridization
    QString saxs_name;            // name for SAXS coefficients
    float   scat_len;             // Scattering length in H20 (*10^-12 cm)
    int     exch_prot;            // Number of exchangable protons
-   int     num_elect;            // Number of electrons
+   double  num_elect;            // Number of electrons
    int     hydrogens;            // Number of hydrogens
-   float   protons;              // Total # of protons
+   double  protons;              // Total # of protons
 };
 
 struct atom
